archive/setup_files/requirements_gui.txt
# Smart Surveillance System - Dependencies

# Core Computer Vision
opencv-python==4.8.1.78
ultralytics==8.0.196

# GUI Framework
tkinter  # Usually comes with Python, but may need manual installation on some systems
Pillow==10.0.1

# Facial Expression Recognition
fer==22.5.1
tensorflow==2.13.0

# Email functionality
smtplib  # Built-in Python module
email     # Built-in Python module

# Video processing (for violence detection)
torch==2.0.1
torchvision==0.15.2
pytorchvideo==0.1.5

# Utility libraries
numpy==1.24.3
threading  # Built-in Python module
queue      # Built-in Python module
datetime   # Built-in Python module

# Optional: For advanced logging
loguru==0.7.2

# Optional: For system notifications
plyer==2.1.0